+++ back-end/src/main/java/com/feedhanjum/back_end/feedback/controller/dto/request/FrequentFeedbackRequestForApiRequest.java
@@ -1,7 +1,7 @@
 package com.feedhanjum.back_end.feedback.controller.dto.request;
 
 import com.feedhanjum.back_end.core.constraints.ByteLength;
-import com.feedhanjum.back_end.team.domain.FrequentFeedbackRequest;
+import com.feedhanjum.back_end.feedback.domain.FrequentFeedbackRequest;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
