+++ back-end/src/main/java/com/feedhanjum/back_end/feedback/controller/dto/response/RegularFeedbackRequestForApiResponse.java
@@ -1,7 +1,7 @@
 package com.feedhanjum.back_end.feedback.controller.dto.response;
 
+import com.feedhanjum.back_end.feedback.domain.RegularFeedbackRequest;
 import com.feedhanjum.back_end.member.controller.dto.MemberDto;
-import com.feedhanjum.back_end.schedule.domain.RegularFeedbackRequest;
 
 import java.time.LocalDateTime;
