+++ back-end/src/main/java/com/feedhanjum/back_end/feedback/repository/FrequentFeedbackRequestRepository.java
@@ -1,6 +1,6 @@
-package com.feedhanjum.back_end.team.repository;
+package com.feedhanjum.back_end.feedback.repository;
 
-import com.feedhanjum.back_end.team.domain.FrequentFeedbackRequest;
+import com.feedhanjum.back_end.feedback.domain.FrequentFeedbackRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FrequentFeedbackRequestRepository extends JpaRepository<FrequentFeedbackRequest, Long> {
