+++ back-end/src/main/java/com/feedhanjum/back_end/feedback/repository/RegularFeedbackRequestRepository.java
@@ -1,7 +1,7 @@
-package com.feedhanjum.back_end.schedule.repository;
+package com.feedhanjum.back_end.feedback.repository;
 
+import com.feedhanjum.back_end.feedback.domain.RegularFeedbackRequest;
 import com.feedhanjum.back_end.member.domain.Member;
-import com.feedhanjum.back_end.schedule.domain.RegularFeedbackRequest;
 import com.feedhanjum.back_end.schedule.domain.ScheduleMember;
 import org.springframework.data.jpa.repository.JpaRepository;
