+++ back-end/src/main/java/com/feedhanjum/back_end/feedback/service/FeedbackQueryService.java
@@ -5,14 +5,14 @@ import com.feedhanjum.back_end.feedback.controller.dto.response.FrequentFeedback
 import com.feedhanjum.back_end.feedback.controller.dto.response.RegularFeedbackRequestForApiResponse;
 import com.feedhanjum.back_end.feedback.domain.Feedback;
 import com.feedhanjum.back_end.feedback.domain.FeedbackReport;
+import com.feedhanjum.back_end.feedback.domain.FrequentFeedbackRequest;
+import com.feedhanjum.back_end.feedback.domain.RegularFeedbackRequest;
 import com.feedhanjum.back_end.feedback.repository.FeedbackQueryRepository;
+import com.feedhanjum.back_end.feedback.repository.FrequentFeedbackRequestQueryRepository;
+import com.feedhanjum.back_end.feedback.repository.RegularFeedbackRequestQueryRepository;
 import com.feedhanjum.back_end.feedback.service.dto.ReceivedFeedbackDto;
 import com.feedhanjum.back_end.feedback.service.dto.SentFeedbackDto;
 import com.feedhanjum.back_end.member.repository.MemberRepository;
-import com.feedhanjum.back_end.schedule.domain.RegularFeedbackRequest;
-import com.feedhanjum.back_end.schedule.repository.RegularFeedbackRequestQueryRepository;
-import com.feedhanjum.back_end.team.domain.FrequentFeedbackRequest;
-import com.feedhanjum.back_end.team.repository.FrequentFeedbackRequestQueryRepository;
 import com.feedhanjum.back_end.team.repository.TeamRepository;
 import jakarta.annotation.Nullable;
 import jakarta.persistence.EntityNotFoundException;
