+++ back-end/src/main/java/com/feedhanjum/back_end/schedule/service/ScheduleService.java
@@ -3,6 +3,7 @@ package com.feedhanjum.back_end.schedule.service;
 import com.feedhanjum.back_end.core.domain.JobRecord;
 import com.feedhanjum.back_end.core.repository.JobRecordRepository;
 import com.feedhanjum.back_end.event.EventPublisher;
+import com.feedhanjum.back_end.feedback.repository.RegularFeedbackRequestQueryRepository;
 import com.feedhanjum.back_end.member.domain.Member;
 import com.feedhanjum.back_end.member.repository.MemberQueryRepository;
 import com.feedhanjum.back_end.member.repository.MemberRepository;
@@ -13,7 +14,6 @@ import com.feedhanjum.back_end.schedule.event.ScheduleEndedEvent;
 import com.feedhanjum.back_end.schedule.exception.ScheduleAlreadyExistException;
 import com.feedhanjum.back_end.schedule.exception.ScheduleIsAlreadyEndException;
 import com.feedhanjum.back_end.schedule.exception.ScheduleMembershipNotFoundException;
-import com.feedhanjum.back_end.schedule.repository.RegularFeedbackRequestQueryRepository;
 import com.feedhanjum.back_end.schedule.repository.ScheduleMemberRepository;
 import com.feedhanjum.back_end.schedule.repository.ScheduleQueryRepository;
 import com.feedhanjum.back_end.schedule.repository.ScheduleRepository;
